+++ project/Common.scala
@@ -21,7 +21,6 @@
 import com.typesafe.sbt.GitPlugin.autoImport._
 import com.typesafe.sbt.site.SiteScaladocPlugin
 import com.typesafe.sbt.site.sphinx.SphinxPlugin
-import com.typesafe.sbt.site.sphinx.SphinxPlugin.autoImport.Sphinx
 import com.typesafe.sbt.{GitBranchPrompt, GitVersioning}
 import org.scalafmt.sbt.ScalafmtPlugin.autoImport.scalafmtOnCompile
 import sbt.Keys._
@@ -97,9 +96,9 @@ object Common {
 object Versions {
   val sparkAvro = "4.0.0"
   val curator = "2.6.0"
-  val spark2d4 = "2.4.7"
-  val spark3d0 = "3.1.1"
-  val scalatest = "3.2.8"
+  val spark2d4 = "2.4.8"
+  val spark3d0 = "3.1.2"
+  val scalatest = "3.2.9"
   val sparkXML = "0.12.0"
   val springBoot = "2.0.6.RELEASE"
   val typesafeConfig = "1.4.1"
@@ -116,18 +115,18 @@ object Versions {
   val esSpark212 = "7.12.1"
   val scopt = "4.0.1"
   val sttp = "1.7.2"
-  val gcs = "1.9.4-hadoop3"
+  val gcs = "hadoop3-2.2.1"
   val hadoopbq = "hadoop3-1.0.0"
   val bq = "1.120.0"
   val hadoop = "3.3.0"
   val h2 = "1.4.200" // Test only
-  val poi = "5.0.0"
+  val poi = "4.1.2"
   val scalate = "1.9.6"
   val akkaHttp = "10.1.14"
 //  val akkaStream = "2.6.12"
   val akkaStream = "2.5.32"
   val kafkaClients = "2.8.0"
-  val testContainers = "0.39.3"
+  val testContainers = "0.39.5"
   val jsqlparser = "4.0"
 }