+++ src/main/scala/com/ebiznext/comet/job/index/kafkaload/KafkaJob.scala
@@ -3,9 +3,11 @@ package com.ebiznext.comet.job.index.kafkaload
 import com.ebiznext.comet.config.Settings
 import com.ebiznext.comet.utils.kafka.KafkaClient
 import com.ebiznext.comet.utils.{JobResult, SparkJob, SparkJobResult, Utils}
+import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.streaming.Trigger
 
+import java.time.LocalDateTime
 import scala.util.Try
 
 class KafkaJob(
@@ -19,8 +21,8 @@ class KafkaJob(
   def offload(): Try[SparkJobResult] = {
     Try {
       if (!kafkaJobConfig.streaming) {
-        Utils.withResources(new KafkaClient(settings.comet.kafka)) { kafkaUtils =>
-          val (df, offsets) = kafkaUtils
+        Utils.withResources(new KafkaClient(settings.comet.kafka)) { kafkaClient =>
+          val (df, offsets) = kafkaClient
             .consumeTopicBatch(
               kafkaJobConfig.topicConfigName,
               session,
@@ -28,13 +30,38 @@ class KafkaJob(
             )
 
           val transformedDF = transfom(df)
-
-          transformedDF.write
+          val finalDF =
+            kafkaJobConfig.coalesce match {
+              case None    => transformedDF
+              case Some(x) => transformedDF.coalesce(x)
+            }
+          finalDF.write
             .mode(kafkaJobConfig.mode)
             .format(kafkaJobConfig.format)
             .options(kafkaJobConfig.writeOptions)
             .save(kafkaJobConfig.path)
-          kafkaUtils.topicSaveOffsets(
+
+          kafkaJobConfig.coalesce match {
+            case Some(1) =>
+              val extension = kafkaJobConfig.format
+              val targetPath = new Path(kafkaJobConfig.path)
+              val singleFile = settings.storageHandler
+                .list(
+                  targetPath,
+                  s".$extension",
+                  LocalDateTime.MIN,
+                  recursive = false
+                )
+                .head
+              val tmpPath = new Path(targetPath.toString + ".tmp")
+              if (settings.storageHandler.move(singleFile, tmpPath)) {
+                settings.storageHandler.delete(targetPath)
+                settings.storageHandler.move(tmpPath, targetPath)
+              }
+            case _ =>
+          }
+
+          kafkaClient.topicSaveOffsets(
             kafkaJobConfig.topicConfigName,
             topicConfig.accessOptions,
             offsets
@@ -42,8 +69,8 @@ class KafkaJob(
           SparkJobResult(Some(transformedDF))
         }
       } else {
-        Utils.withResources(new KafkaClient(settings.comet.kafka)) { kafkaUtils =>
-          val df = kafkaUtils
+        Utils.withResources(new KafkaClient(settings.comet.kafka)) { kafkaClient =>
+          val df = kafkaClient
             .consumeTopicStreaming(
               session,
               topicConfig
@@ -84,11 +111,11 @@ class KafkaJob(
 
   def load(): Try[SparkJobResult] = {
     Try {
-      Utils.withResources(new KafkaClient(settings.comet.kafka)) { kafkaUtils =>
+      Utils.withResources(new KafkaClient(settings.comet.kafka)) { kafkaClient =>
         val df = session.read.format(kafkaJobConfig.format).load(kafkaJobConfig.path.split(','): _*)
         val transformedDF = transfom(df)
 
-        kafkaUtils.sinkToTopic(
+        kafkaClient.sinkToTopic(
           topicConfig,
           transformedDF
         )