+++ src/main/scala/com/ebiznext/comet/job/index/kafkaload/KafkaJobConfig.scala
@@ -18,7 +18,8 @@ case class KafkaJobConfig(
   streamingTrigger: String = "Once",
   streamingTriggerOption: String = "",
   streamingWritePartitionBy: Seq[String] = Nil,
-  streamingWriteToTable: Boolean = false
+  streamingWriteToTable: Boolean = false,
+  coalesce: Option[Int] = None
 ) {
 
   val transformInstance: Option[DataFrameTransform] = {
@@ -38,13 +39,33 @@ object KafkaJobConfig extends CliConfig[KafkaJobConfig] {
     OParser.sequence(
       programName("comet kafkaload"),
       head("comet", "kafkaload", "[options]"),
-      note(""),
+      note("""
+          |Two modes are available : The batch mode and the streaming mode.
+          |
+          |### Batch mode
+          |In batch mode, you start the kafka (off)loader regurarly and the last consumed offset 
+          |will be stored in the `comet_offsets` topic config 
+          |(see [reference-kafka.conf](https://github.com/ebiznext/comet-data-pipeline/blob/master/src/main/resources/reference-kafka.conf#L22) for an example).
+          |
+          |When offloading data from kafka to a file, you may ask to coalesce the result to a specific number of files / partitions.
+          |If you ask to coalesce to a single partition, the offloader will store the data in the exact filename you provided in the path
+          |argument.
+          |
+          |The figure below describes the batch offloading process
+          |![](/img/cli/kafka-offload.png)
+          |
+          |
+          |### Streaming mode
+          |
+          |In this mode, te program keep running and you the comet_offsets topic is not used. The (off)loader will use a consumer group id 
+          |you specify in the access options of the topic configuration you are dealing with.
+          |""".stripMargin),
       opt[String]("topic")
         .action((x, c) => c.copy(topicConfigName = x))
         .text("Topic Name declared in reference.conf file")
         .required(),
       opt[String]("format")
-        .action((x, c) => c.copy(topicConfigName = x))
+        .action((x, c) => c.copy(format = x))
         .text("Read/Write format eq : parquet, json, csv ... Default to parquet.")
         .optional(),
       opt[String]("path")
@@ -54,13 +75,19 @@ object KafkaJobConfig extends CliConfig[KafkaJobConfig] {
       opt[String]("mode")
         .action((x, c) => c.copy(mode = SaveMode.valueOf(x)))
         .text(
-          "When offload is true, describes who data should be stored on disk. Ignored if offload is false."
+          "When offload is true, describes how data should be stored on disk. Ignored if offload is false."
         )
         .required(),
       opt[Map[String, String]]("write-options")
         .action((x, c) => c.copy(writeOptions = x))
         .text(
           "Options to pass to Spark Writer"
+        )
+        .optional(),
+      opt[Int]("coalesce")
+        .action((x, c) => c.copy(coalesce = Some(x)))
+        .text(
+          "Should we coalesce the resulting dataframe"
         )
         .optional(),
       opt[String]("transform")
@@ -76,20 +103,20 @@ object KafkaJobConfig extends CliConfig[KafkaJobConfig] {
           opt[Unit]("stream")
             .action((_, c) => c.copy(streaming = true))
             .text(
-              "If true, kafka topic is offloaded to path, else data contained in path is stored in the kafka topic"
+              "Should we use streaming mode ?"
             )
             .optional()
             .children(
               opt[String]("streaming-format")
                 .action((x, c) => c.copy(streamingWriteFormat = x))
                 .text(
-                  "If true, kafka topic is offloaded to path, else data contained in path is stored in the kafka topic"
+                  "Streaming format eq. kafka, console ..."
                 )
                 .required(),
               opt[String]("streaming-output-mode")
                 .action((x, c) => c.copy(streamingWriteMode = x))
                 .text(
-                  "If true, kafka topic is offloaded to path, else data contained in path is stored in the kafka topic"
+                  "Output mode : eq. append ... "
                 )
                 .required(),
               opt[String]("streaming-trigger")