+++ src/main/scala/com/ebiznext/comet/job/infer/InferSchemaJob.scala
@@ -24,9 +24,10 @@ import com.ebiznext.comet.config.{Settings, SparkEnv}
 import com.ebiznext.comet.schema.handlers.InferSchemaHandler
 import com.ebiznext.comet.schema.model.{Attribute, Domain}
 import org.apache.hadoop.fs.Path
-import org.apache.spark.sql.{DataFrame, Dataset, SparkSession}
+import org.apache.spark.sql.{DataFrame, Dataset, Encoders, SparkSession}
 
 import java.util.regex.Pattern
+import scala.io.Source
 import scala.util.Try
 
 /** *
@@ -71,54 +72,45 @@ class InferSchemaJob(implicit settings: Settings) {
   }
 
   /** Get format file by using the first and the last line of the dataset
-    * We use mapPartitionsWithIndex to retrieve these informations to make sure that the first line really corresponds to the first line (same for the last)
+    * We use mapPartitionsWithIndex to retrieve these information to make sure that the first line really corresponds to the first line (same for the last)
     *
-    * @param datasetInit : created dataset without specifying format
+    * @param lines : list of lines read from file
     * @return
     */
-  def getFormatFile(datasetInit: Dataset[String]): String = {
-    val rddDatasetInit = datasetInit.rdd
-    val lastPartitionNo = rddDatasetInit.getNumPartitions - 1
+  def getFormatFile(lines: List[String]): String = {
+    val firstLine = lines.head
+    val lastLine = lines.last
 
-    //Retrieve the first and the last line of a dataset
-    val partitionWithIndex = rddDatasetInit.mapPartitionsWithIndex { (index, iterator) =>
-      {
-        val iteratorList = iterator.toList
-        //The first line is stored into the 0th partition
-        //Check if data is stored on the same partition (if true return directly the first and the last line)
-        (index, lastPartitionNo) match {
-          case (0, 0) => Iterator(iteratorList.take(1).head, iteratorList.reverse.take(1).last)
-          case (0, _) => iteratorList.take(1).iterator
-          //The last line is stored into the last partition
-          case (i, l) if i == l => iteratorList.reverse.take(1).iterator
-          case (_, _)           => Iterator()
-        }
-      }
+    val jsonRegexStart = """\{.*""".r
+    val jsonArrayRegexStart = """\[.*""".r
+
+    val jsonRegexEnd = """.*\}""".r
+    val jsonArrayRegexEnd = """.*\]""".r
+
+    (firstLine, lastLine) match {
+      case (jsonRegexStart(), jsonRegexEnd())           => "JSON"
+      case (jsonArrayRegexStart(), jsonArrayRegexEnd()) => "ARRAY_JSON"
+      case _                                            => "DSV"
     }
-
-    val firstLine = partitionWithIndex.first
-    val lastLine = partitionWithIndex.collect().last
-
-    if (firstLine.startsWith("{") && firstLine.endsWith("}")) "JSON"
-    else if (firstLine.startsWith("[") && lastLine.endsWith("]")) "ARRAY_JSON"
-    else "DSV"
-
   }
 
   /** Get separator file by taking the character that appears the most in 10 lines of the dataset
     *
-    * @param datasetInit : created dataset without specifying format
+    * @param lines : list of lines read from file
     * @return the file separator
     */
-  def getSeparator(datasetInit: Dataset[String]): String = {
-    val (separator, _) = session.sparkContext
-      .parallelize(datasetInit.take(10))
-      .map(x => x.replaceAll("[A-Za-z0-9 \"'()@?!éèîàÀÉÈç+]", ""))
-      .flatMap(_.toCharArray)
-      .map(w => (w, 1))
-      .reduceByKey(_ + _)
-      .max
-    separator.toString
+  def getSeparator(lines: List[String]): String = {
+    val firstLine = lines.head
+    val separator =
+      firstLine
+        .replaceAll("[A-Za-z0-9 \"'()@?!éèîàÀÉÈç+]", "")
+        .toCharArray
+        .map((_, 1))
+        .groupBy(_._1)
+        .mapValues(_.size)
+        .toList
+        .maxBy(_._2)
+    separator._1.toString
   }
 
   /** Get domain directory name
@@ -141,32 +133,40 @@ class InferSchemaJob(implicit settings: Settings) {
 
   /** Create the dataframe with its associated format
     *
-    * @param datasetInit : created dataset without specifying format
+    * @param lines : list of lines read from file
     * @param path        : file path
     * @return
     */
   def createDataFrameWithFormat(
-    datasetInit: Dataset[String],
-    path: Path,
+    lines: List[String],
+    dataPath: String,
     header: Boolean
   ): DataFrame = {
-    val formatFile = getFormatFile(datasetInit)
-
+    val formatFile = getFormatFile(lines)
     formatFile match {
-      case "JSON" | "ARRAY_JSON" =>
+      case "ARRAY_JSON" =>
+        val jsonRDD =
+          session.sparkContext.wholeTextFiles(dataPath).map { case (_, content) =>
+            content
+          }
+        session.read
+          .option("inferSchema", value = true)
+          .json(session.createDataset(jsonRDD)(Encoders.STRING))
+
+      case "JSON" =>
         session.read
           .format("json")
           .option("inferSchema", value = true)
-          .load(path.toString)
+          .load(dataPath)
 
       case "DSV" =>
         session.read
           .format("com.databricks.spark.csv")
           .option("header", header)
           .option("inferSchema", value = true)
-          .option("delimiter", getSeparator(datasetInit))
+          .option("delimiter", getSeparator(lines))
           .option("parserLib", "UNIVOCITY")
-          .load(path.toString)
+          .load(dataPath)
     }
   }
 
@@ -184,17 +184,17 @@ class InferSchemaJob(implicit settings: Settings) {
     Try {
       val path = new Path(dataPath)
 
-      val datasetWithoutFormat = readFile(path)
+      val lines = Source.fromFile(path.toString).getLines().toList.map(_.trim).filter(_.nonEmpty)
 
-      val dataframeWithFormat = createDataFrameWithFormat(datasetWithoutFormat, path, header)
+      val dataframeWithFormat = createDataFrameWithFormat(lines, dataPath, header)
 
-      val format = Option(getFormatFile(datasetWithoutFormat))
+      val format = getFormatFile(lines)
 
-      val array = format.contains("ARRAY_JSON")
+      val array = format == "ARRAY_JSON"
 
       val withHeader = header
 
-      val separator = getSeparator(datasetWithoutFormat)
+      val separator = getSeparator(lines)
 
       val inferSchema = InferSchemaHandler