+++ src/main/scala/com/ebiznext/comet/schema/model/Domain.scala
@@ -105,7 +105,7 @@ case class Domain(
     *
     * @return the ack attribute or ".ack" by default
     */
-  def getAck(): String = ack.map(ack => if (ack.nonEmpty) "." + ack else ack).getOrElse(".ack")
+  def getAck(): String = ack.map(ack => if (ack.nonEmpty) "." + ack else ack).getOrElse("")
 
   /** Is this Domain valid ? A domain is valid if :
     *   - The domain name is a valid attribute
@@ -151,5 +151,13 @@ case class Domain(
       Left(errorList.toList)
     else
       Right(true)
+  }
+
+  def asDot(includeAllAttrs: Boolean): String = {
+    schemas
+      .map { schema =>
+        schema.asDot(name, includeAllAttrs)
+      }
+      .mkString("\n")
   }
 }