+++ src/main/scala/com/ebiznext/comet/workflow/IngestionWorkflow.scala
@@ -93,9 +93,11 @@ class IngestionWorkflow(
               domain.getExtensions().map(ext => new Path(prefixStr + ext))
           val existRawFile = rawFormats.find(file => storageHandler.exists(file))
           logger.info(s"Found ack file $ackFile")
+          val zipExtensions = List(".tgz", ".gz", ".zip")
+          val isZip = zipExtensions.exists(ext => path.toString.endsWith(ext))
           if (domain.getAck().nonEmpty)
             storageHandler.delete(ackFile)
-          if (existRawFile.isDefined) {
+          if (!isZip && existRawFile.isDefined) {
             existRawFile.foreach { file =>
               logger.info(s"Found raw file $existRawFile")
               storageHandler.mkdirs(tmpDir)
@@ -297,7 +299,7 @@ class IngestionWorkflow(
     val lockPath =
       new Path(settings.comet.lock.path, s"${config.domain}_${config.schema}.lock")
     val locker = new FileLock(lockPath, storageHandler)
-    val waitTimeMillis = settings.comet.lock.ingestionTimeout
+    val waitTimeMillis = settings.comet.lock.timeout
 
     locker.doExclusively(waitTimeMillis) {
       val domainName = config.domain