+++ src/test/scala/com/ebiznext/comet/schema/generator/DDL2YmlSpec.scala
@@ -54,6 +54,45 @@ class DDL2YmlSpec extends TestHelper {
         "long",
         "string"
       )
+      domain.schemas
+        .find(_.name == "TEST_TABLE1")
+        .get
+        .primaryKey
+        .getOrElse(List.empty) should contain("ID")
+    }
+  }
+
+  "JDBCSchemas" should "deserialize corrected" in {
+    new WithSettings() {
+      val imput =
+        """
+          |extract:
+          |  jdbcSchemas:
+          |    - connection: "test-h2" # Connection name as defined in the connections section of the application.conf file
+          |      catalog: "business" # Optional catalog name in the target database
+          |      schema: "public" # Database schema where tables are located
+          |      tables:
+          |        - name: "user"
+          |          columns: # optional list of columns, if not present all columns should be exported.
+          |            - id
+          |            - email
+          |        - name: product # All columns should be exported
+          |        - name: "*" # Ignore any other table spec. Just export all tables
+          |      tableTypes: # One or many of the types below
+          |        - "TABLE"
+          |        - "VIEW"
+          |        - "SYSTEM TABLE"
+          |        - "GLOBAL TEMPORARY"
+          |        - "LOCAL TEMPORARY"
+          |        - "ALIAS"
+          |        - "SYNONYM"
+          |      template: "/my-templates/domain-template.yml" # Metadata to use for the generated YML file.
+          |""".stripMargin
+      val jdbcMapping = File.newTemporaryFile()
+      val outputDir = File.newTemporaryDirectory()
+      jdbcMapping.overwrite(imput)
+      val jdbcSchemas = YamlSerializer.deserializeJDBCSchemas(jdbcMapping)
+      assert(jdbcSchemas.jdbcSchemas.size > 0)
     }
   }
 
@@ -93,6 +132,52 @@ class DDL2YmlSpec extends TestHelper {
       domain.schemas.head.attributes
         .map(_.name) should contain theSameElementsAs Set("ID")
       domain.schemas.head.attributes.map(_.`type`) should contain theSameElementsAs Set("long")
+      domain.schemas.head.primaryKey.getOrElse(List.empty) should contain("ID")
+    }
+  }
+
+  "DDL2Yml with foreign keys" should "detect the foreign keys" in {
+    new WithSettings() {
+      val jdbcOptions = settings.comet.connections("test-h2")
+      val conn = DriverManager.getConnection(
+        jdbcOptions.options("url"),
+        jdbcOptions.options("user"),
+        jdbcOptions.options("password")
+      )
+      val sql: String =
+        """
+          |drop view if exists test_view1;
+          |drop table if exists test_table1;
+          |drop table if exists test_table2;
+          |create table test_table1(ID INT PRIMARY KEY,NAME VARCHAR(500));
+          |create table test_table2(ID INT PRIMARY KEY,TABLE1_ID INT,foreign key (TABLE1_ID) references test_table1(ID));
+          |insert into test_table1 values (1,'A');
+          |insert into test_table2 values (1,1);
+          |insert into test_table2 values (2,1);""".stripMargin
+      val st = conn.createStatement()
+      st.execute(sql)
+      val rs = st.executeQuery("select * from test_table1")
+      rs.next
+      val row1InsertionCheck = (1 == rs.getInt("ID")) && ("A" == rs.getString("NAME"))
+      assert(row1InsertionCheck, "Data not inserted")
+
+      DDL2Yml.run(
+        JDBCSchema("test-h2", None, "PUBLIC", List(JDBCTable("TEST_TABLE2", None))),
+        File("/tmp"),
+        None
+      )
+      val domain = YamlSerializer.deserializeDomain(File("/tmp", "PUBLIC.comet.yml")) match {
+        case Success(domain) => domain
+        case Failure(e)      => throw e
+      }
+      assert(domain.name == "PUBLIC")
+      assert(domain.schemas.size == 1)
+      assert(domain.schemas.head.name == "TEST_TABLE2")
+      domain.schemas.head.attributes
+        .find(_.name == "TABLE1_ID")
+        .get
+        .foreignKey
+        .getOrElse("") should be("PUBLIC.TEST_TABLE1.ID")
     }
   }