+++ src/test/scala/com/ebiznext/comet/schema/handlers/SchemaHandlerSpec.scala
@@ -20,10 +20,12 @@
 
 package com.ebiznext.comet.schema.handlers
 
+import better.files.File
 import com.databricks.spark.xml._
 import com.dimafeng.testcontainers.ElasticsearchContainer
 import com.ebiznext.comet.TestHelper
 import com.ebiznext.comet.config.DatasetArea
+import com.ebiznext.comet.schema.generator.Yml2GraphViz
 import com.ebiznext.comet.schema.model._
 import com.softwaremill.sttp.{HttpURLConnectionBackend, _}
 import com.typesafe.config.{Config, ConfigFactory}
@@ -551,5 +553,102 @@ class SchemaHandlerSpec extends TestHelper {
         write = Some(WriteMode.OVERWRITE)
       )
     }
+    "Exporting domain as Dot" should "create a valid dot file" in {
+      new SpecTrait(
+        domainOrJobFilename = "dream.comet.yml",
+        sourceDomainOrJobPathname = s"/sample/dream/dream.comet.yml",
+        datasetDomainName = "dream",
+        sourceDatasetPathName = "/sample/dream/OneClient_Segmentation_20190101_090800_008.psv"
+      ) {
+        cleanMetadata
+        cleanDatasets
+        val schemaHandler = new SchemaHandler(settings.storageHandler)
+
+        new Yml2GraphViz(schemaHandler).run(Array("--all", "false"))
+
+        val tempFile = File.newTemporaryFile().pathAsString
+        new Yml2GraphViz(schemaHandler).run(
+          Array("--all", "true", "--output", tempFile)
+        )
+        val fileContent = readFileContent(tempFile)
+        val expectedFileContent = loadTextFile("/expected/dot/output.dot")
+        fileContent shouldBe expectedFileContent
+
+        val result = schemaHandler.domains.head.asDot(false)
+        result.trim shouldBe """
+                               |
+                               |dream_segment [label=<
+                               |<table border="0" cellborder="1" cellspacing="0">
+                               |<tr><td port="0" bgcolor="darkgreen"><B><FONT color="white"> segment </FONT></B></td></tr>
+                               |<tr><td port="dream_id"><B> dreamkey:long </B></td></tr>
+                               |</table>>];
+                               |
+                               |
+                               |
+                               |dream_client [label=<
+                               |<table border="0" cellborder="1" cellspacing="0">
+                               |<tr><td port="0" bgcolor="darkgreen"><B><FONT color="white"> client </FONT></B></td></tr>
+                               |<tr><td port="dream_id"><I> dream_id:long </I></td></tr>
+                               |</table>>];
+                               |
+                               |dream_client:dream_id -> dream_segment:0
+                               |
+                               |""".stripMargin.trim
+      }
+    }
+
+    "Ingest Dream Contact CSV with ignore" should "produce file in accepted" in {
+      new SpecTrait(
+        domainOrJobFilename = "dreamignore.comet.yml",
+        sourceDomainOrJobPathname = s"/sample/dream/dreamignore.comet.yml",
+        datasetDomainName = "dreamignore",
+        sourceDatasetPathName = "/sample/dream/OneClient_Contact_20190101_090800_008.psv"
+      ) {
+
+        cleanMetadata
+
+        cleanDatasets
+
+        loadPending
+
+        readFileContent(
+          cometDatasetsPath + s"/archive/$datasetDomainName/OneClient_Contact_20190101_090800_008.psv"
+        ) shouldBe loadTextFile(
+          sourceDatasetPathName
+        )
+
+        //If we run this test alone, we do not have rejected, else we have rejected but not accepted ...
+        Try {
+          printDF(
+            sparkSession.read.parquet(
+              cometDatasetsPath + "/rejected/dream/client"
+            ),
+            "dream/client"
+          )
+        }
+
+        // Accepted should have the same data as input
+        val acceptedDf = sparkSession.read
+          .parquet(
+            cometDatasetsPath + s"/accepted/$datasetDomainName/client/${getTodayPartitionPath}"
+          )
+          // Timezone Problem
+          .drop("customer_creation_date")
+
+        val expectedAccepted =
+          sparkSession.read
+            .schema(acceptedDf.schema)
+            .json(getResPath("/expected/datasets/accepted/dream/clientignore.json"))
+            // Timezone Problem
+            .drop("customer_creation_date")
+            .withColumn("truncated_zip_code", substring(col("zip_code"), 0, 3))
+            .withColumn("source_file_name", lit("OneClient_Contact_20190101_090800_008.psv"))
+
+        acceptedDf.columns.size shouldBe expectedAccepted.columns.size
+        acceptedDf.except(expectedAccepted).count() shouldBe 0
+      }
+
+    }
+
   }
 }